# src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: Spring eprescribing
  description: Spring ePrescribing Application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  version: '1.0'
servers:
  - url: http://localhost:9966/eprescribing/api
...
paths:
...
  /specialties:
    get:
      tags:
        - specialty
      operationId: listSpecialties
      summary: Lists specialties
      description: Returns an array of specialty .
      responses:
        200:
          description: Specialties found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Specialty'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
    post:
      tags:
        - specialty
      operationId: addSpecialty
      summary: Create a specialty
      description: Creates a specialty .
      requestBody:
        description: The specialty
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Specialty'
        required: true
      responses:
        200:
          description: Specialty created successfully.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Specialty'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
        404:
          description: Specialty not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
...
  /vets:
    get:
      tags:
        - vet
      operationId: listVets
      summary: Lists vets
      description: Returns an array of vets.
      responses:
        200:
          description: Vets found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Vet'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'

    post:
      tags:
        - vet
      operationId: addVet
      summary: Create a Vet
      description: Creates a vet .
      requestBody:
        description: The vet
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vet'
        required: true
      responses:
        200:
          description: Vet created successfully.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vet'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
        404:
          description: Vet not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
  /vets/{vetId}:
    get:
      tags:
        - vet
      operationId: getVet
      summary: Get a vet by ID
      description: Returns the vet or a 404 error.
      parameters:
        - name: vetId
          in: path
          description: The ID of the vet.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      responses:
        200:
          description: Vet details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vet'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
        404:
          description: Vet not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
    put:
      tags:
        - vet
      operationId: updateVet
      summary: Update a vet  by ID
      description: Returns the vet or a 404 error.
      parameters:
        - name: vetId
          in: path
          description: The ID of the vet.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      requestBody:
        description: The vet
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Vet'
        required: true
      responses:
        200:
          description: Pet type details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vet'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
        404:
          description: Vet not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
    delete:
      tags:
        - vet
      operationId: deleteVet
      summary: Delete a vet by ID
      description: Returns the vet or a 404 error.
      parameters:
        - name: vetId
          in: path
          description: The ID of the vet.
          required: true
          schema:
            type: integer
            format: int32
            minimum: 0
            example: 1
      responses:
        200:
          description: Vet details found and returned.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vet'
        304:
          description: Not modified.
          headers:
            ETag:
              description: An ID for this version of the response.
              schema:
                type: string
        400:
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
        404:
          description: Vet not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
        500:
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestError'
...
components:
  schemas:
    RestError:
      title: REST Error
      description: The schema for all error responses.
      type: object
      properties:
        status:
          title: Status
          description: The HTTP status code.
          type: integer
          format: int32
          example: 400
          readOnly: true
        error:
          title: Error
          description: The short error message.
          type: string
          example: Bad Request
          readOnly: true
        path:
          title: Path
          description: The path of the URL for this request.
          type: string
          format: uri
          example: '/api/owners'
          readOnly: true
        timestamp:
          title: Timestamp
          description: The time the error occured.
          type: string
          format: date-time
          example: '2019-08-21T21:41:46.158+0000'
          readOnly: true
        message:
          title: Message
          description: The long error message.
          type: string
          example: 'Request failed schema validation'
          readOnly: true
        schemaValidationErrors:
          title: Schema validation errors
          description: Validation errors against the OpenAPI schema.
          type: array
          items:
            $ref: '#/components/schemas/ValidationMessage'
        trace:
          title: Trace
          description: The stacktrace for this error.
          type: string
          example: 'com.atlassian.oai.validator.springmvc.InvalidRequestException: ...'
          readOnly: true
      required:
        - status
        - error
        - path
        - timestamp
        - message
        - schemaValidationErrors
    ValidationMessage:
      title: Validation message
      description: Messages describing a validation error.
      type: object
      properties:
        message:
          title: Message
          description: The valiation message.
          type: string
          example: "[Path '/lastName'] Instance type (null) does not match any allowed primitive type (allowed: [\"string\"])"
          readOnly: true
      required:
        - message
      additionalProperties: true
    Specialty:
      title: Specialty
      description: Fields of specialty of vets.
      type: object
      properties:
        id:
          title: ID
          description: The ID of the specialty.
          type: integer
          format: int32
          minimum: 0
          example: 1
          readOnly: true
        name:
          title: Name
          description: The name of the specialty.
          type: string
          maxLength: 80
          minLength: 1
          example: radiology
      required:
        - id
        - name
...
    VetFields:
      title: VetFields
      description: Editable fields of a veterinarian.
      type: object
      properties:
        firstName:
          title: First name
          description: The first name of the vet.
          type: string
          minLength: 1
          maxLength: 30
          pattern: '^[a-zA-Z]*$'
          example: 'James'
        lastName:
          title: Last name
          description: The last name of the vet.
          type: string
          minLength: 1
          maxLength: 30
          pattern: '^[a-zA-Z]*$'
          example: 'Carter'
        specialties:
          title: Specialties
          description: The specialties of the vet.
          type: array
          items:
            $ref: '#/components/schemas/Specialty'
      required:
        - firstName
        - lastName
        - specialties
    Vet:
      title: Vet
      description: A veterinarian.
      allOf:
        - $ref: '#/components/schemas/VetFields'
        - type: object
          properties:
            id:
              title: ID
              description: The ID of the vet.
              type: integer
              format: int32
              minimum: 0
              example: 1
              readOnly: true
      required:
        - id
        - firstName
        - lastName
        - specialties
...